[REFACTOR] 쿠폰정책 기존타입 bigdicimal로 변경

+++ src/main/java/com/nhnacademy/illuwa/domain/coupons/dto/couponPolicy/CouponPolicyCreateRequest.java
@@ -7,6 +7,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
@@ -17,12 +19,12 @@ public class CouponPolicyCreateRequest {
     private String code;
 
     @NotNull
-    private int minOrderAmount; // 최소 주문 금액
+    private BigDecimal minOrderAmount; // 최소 주문 금액
 
-    private Integer discountAmount; // 할인 금액
+    private BigDecimal discountAmount; // 할인 금액
 
-    private Integer discountPercent; // 할인 퍼센트
+    private BigDecimal discountPercent; // 할인 퍼센트
 
-    private Integer maxDiscountAmount; // 최대 할인 금액
+    private BigDecimal maxDiscountAmount; // 최대 할인 금액
 
 }
